services/worker/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0.post1
httpx==0.25.1
pyodm==1.5.12
rasterio==1.3.9
numpy==1.26.2
torch==2.1.1
opencv-python-headless==4.8.1.78
python-multipart==0.0.6
psycopg2-binary==2.9.9
redis==5.0.1
